Add LspReference{Text,Read,Write}
Part of Neovim's LSP implementation, they are used when the LSP is able
to present highlights to the requested reference in the file.

Typically editors use this when the cursor idles on something like a
variable to help show where else that variable is used.

The different groups help to distinguish between:

  - when a reference is set, i.e. `LspReferenceWrite`
  - when a reference is used, i.e. `LspReferenceRead`
  - when a reference is neither the above or perhaps just a literal,
    i.e. `LspReferenceRead`